Created Sign up form and Login Form. Fixed LAnding page Slogan when button is pressed. Fixed Login and Create button pop up interference.
