+++ client/src/App.js
@@ -1,8 +1,11 @@
 // import MainPage from "./components/pages/MainPage";
 import LandingPage from "./components/pages/LandingPage";
-
 function App() {
-  return <LandingPage />;
+  return (
+    <>
+      <LandingPage />
+    </>
+  );
 }
 
 export default App;
