+++ client/src/components/css/landingpage.css
@@ -6,7 +6,14 @@
   position: absolute;
   object-fit: cover;
 }
-
+.bg-img {
+  background-position: center;
+  background-image: url("../../images/landing.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  height: 100vh;
+  
+}
 .btn-div,
 #small-device-btn {
   padding: min(2em, 3em);
